[Add] excludePathPatterns 추가
